Remove redux usage
