+++ src/components/Collections/CollectionDialog.js
@@ -1,24 +1,42 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import axios from 'axios';
-import { useDispatch } from 'react-redux';
 import { Grid } from '@mui/material';
-import { addUserCollection } from '../../redux/actions';
+import { useMutation } from '@tanstack/react-query';
 import { YupInput, LoaderButton } from '../Miscellaneous';
 import YupDialog from '../Miscellaneous/YupDialog';
-import { apiBaseUrl } from '../../config';
 import { useAuth } from '../../contexts/AuthContext';
 import useToast from '../../hooks/useToast';
+import callYupApi from '../../apis/base_api';
+import { queryClient } from '../../config/react-query';
+import { REACT_QUERY_KEYS } from '../../constants/enum';
 
 const TITLE_LIMIT = 30;
 const DESC_LIMIT = 140;
 
 function CollectionDialog({ postid, dialogOpen, handleDialogClose }) {
-  const dispatch = useDispatch();
   const { authInfo, userId } = useAuth();
   const [description, setDescription] = useState('');
   const [name, setName] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
+  const { isLoading, mutate } = useMutation(
+    (data) => callYupApi({
+      url: '/collections',
+      method: 'POST',
+      data
+    }),
+    {
+      onSuccess: (data) => {
+        queryClient.setQueryData(
+          [REACT_QUERY_KEYS.USER_COLLECTIONS, userId],
+          (oldData) => oldData ? [data, ...oldData] : undefined
+        );
+
+        toastSuccess(`Successfully created ${data.name}`);
+
+        handleDialogClose();
+      },
+      onError: () => toastError('There was a problem creating your collection')
+    }
+  );
 
   const { toastSuccess, toastError } = useToast();
 
@@ -28,26 +46,15 @@ function CollectionDialog({ postid, dialogOpen, handleDialogClose }) {
     if (e.key === 'Enter' && !!name) handleCreateNewCollection();
   };
   const handleCreateNewCollection = async () => {
-    try {
-      if (isLoading) return;
-      setIsLoading(true);
-      const postId = postid === 'routeFromUrl' ? undefined : postid;
-      const params = {
-        name,
-        description,
-        postId,
-        eosname: authInfo.eosname,
-        ...authInfo
-      };
-      const { data } = await axios.post(`${apiBaseUrl}/collections`, params);
-      dispatch(addUserCollection(userId, data));
-      toastSuccess(`Succesfully created ${name}`);
-      handleDialogClose();
-      setIsLoading(false);
-    } catch (err) {
-      setIsLoading(false);
-      toastError(`There was a problem creating your collection`);
-    }
+    const postId = postid === 'routeFromUrl' ? undefined : postid;
+
+    mutate({
+      name,
+      description,
+      postId,
+      eosname: authInfo.eosname,
+      ...authInfo
+    });
   };
 
   return (
